requirements.txt
google-cloud-aiplatform>=1.38.0
numpy>=2.1.1
qdrant-client>=1.9.0
python-dotenv>=1.0.1
pytest>=8.0.0
fastapi>=0.100.0
uvicorn>=0.23.0